Split navbar into sidebar, header and view
