+++ src/components/Header/Header.jsx
@@ -1,0 +1,37 @@
+import React, { Component } from 'react';
+
+//import pictures for navbar and sidebar
+import Logo from '../../../assets-src/Navbar/inMarketlogo.png';
+
+//import stylesheet
+import './Header.css';
+
+//import Framework 7 files
+import {
+  Navbar,
+  List,
+  ListItem,
+  NavTitle,
+  NavRight,
+  Link,
+} from 'framework7-react';
+
+/* Contains both the header and link to sidebar */
+class Header extends Component {
+  render() {
+    return (
+        <Navbar className="navbar" sliding={false}>
+          <div className="navbarcontent">
+                <img className="navbarlogo" src={Logo} />
+                <NavTitle className="navbartitle">NAVN</NavTitle>
+                <NavTitle className="navbartitle">NOTIFIKASJONER</NavTitle>
+                <NavRight className="navright">
+                  <Link panelOpen="right"><NavTitle className="navbartitle">MENU</NavTitle></Link>
+                </NavRight>
+          </div>
+        </Navbar>
+    );
+  }
+}
+
+export default Header;
